DB is being made correctly, need to update Controllers for new relations

+++ MotoDex/MotoDex/Models/Tyre.cs
@@ -10,7 +10,7 @@ namespace MotoDex.Models
         public int Id { get; set; }
         public string Make { get; set; }
         public string Model { get; set; }
-        public int TyreWidth { get; set; }
+        public float TyreWidth { get; set; }
         public int HeightAspect { get; set; }
         public int RimSize { get; set; }
